feat: Implement booking system with venue selection, date management, and email confirmation
- Added booking.js for managing booking state, available dates, and session modals.
- Created config.js to store configuration for venues, email settings, and instructions.
- Developed email.js to handle confirmation email sending with EmailJS integration.
- Introduced ui.js for UI management, including smooth scrolling, form validation, and loading messages.
- Added utility functions in utils.js for email validation and unique booking ID generation.
