+++ index.html
@@ -267,7 +267,7 @@
     <script type="text/javascript" src="https://cdn.jsdelivr.net/npm/@emailjs/browser@4/dist/email.min.js"></script>
     
     <!-- JavaScript -->
-    <script src="script.js"></script>
+    <script type="module" src="script.js"></script>
 
     <!-- JSON-LD Structured Data for SEO -->
     <script type="application/ld+json">
