+++ js/utils.js
@@ -1,0 +1,40 @@
+// Mummy's Messy Makers - Utility Functions
+
+export const utils = {
+    debounce: function(func, wait) {
+        let timeout;
+        return function executedFunction(...args) {
+            const later = () => {
+                clearTimeout(timeout);
+                func(...args);
+            };
+            clearTimeout(timeout);
+            timeout = setTimeout(later, wait);
+        };
+    },
+
+    throttle: function(func, limit) {
+        let inThrottle;
+        return function() {
+            const args = arguments;
+            const context = this;
+            if (!inThrottle) {
+                func.apply(context, args);
+                inThrottle = true;
+                setTimeout(() => inThrottle = false, limit);
+            }
+        };
+    }
+};
+
+// Email validation helper
+export function isValidEmail(email) {
+    return email.includes('@') && email.includes('.');
+}
+
+// Generate unique booking ID
+export function generateBookingId() {
+    const timestamp = Date.now().toString(36);
+    const random = Math.random().toString(36).substring(2, 7);
+    return `MMM-${timestamp}-${random}`.toUpperCase();
+}
